backend/requirements.txt
pytest
pytest-asyncio
pydantic
PyYAML
fastapi[standard]
jinja2
pydantic-settings
groq
python-dotenv
